tools/nasqueron-reports/requirements.txt
hvac~=2.3.0
mysql-connector-python~=9.4.0
PyYAML~=6.0.2
sqlparse~=0.5.3
